tests/wpt/tests/workers/interfaces/DedicatedWorkerGlobalScope/postMessage/second-argument-undefined.html

<!doctype html>
<title>Using undefined in postMessage's second argument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new Worker('second-argument-undefined.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 1);
    this.done();
  });
});
</script>
